requirement.txt
langchain

langchain-core
langchain-community
langchain-openai

pypdf
python-dotenv
openai
qdrant-client
langchain-qdrant
